Update flashy for scarthgap image naming
With the latest scarthgap update for AsteroidOS, the image name has
changed and now includes ".rootfs" so the flashy utility has now been
updated to match that.  This means that it is no longer able to flash
images created using the previous naming scheme, but it's a trivial
change to either refert this change or simply rename the image file.

Signed-off-by: Ed Beroset <beroset@ieee.org>