Use better method for finding host's IP address
Int the setRouting() function, the HOSTNAME command was used to get the
IP address of the host, but this method doesn't work for the emulator
target.  This fixes that by running the "who" command on the watch and
then picking off the last field (host's IP address) from the response
via awk.

Signed-off-by: Ed Beroset <beroset@ieee.org>